[#2387] - Reubica los handles por identidad bajo el corpus que la regla alcanza
Los agregadores exponían, además de sus colecciones y selectores, un handle por obra y por
colección. Ese nombre era la vía por la que se alcanzaba una pieza puntual del corpus desde fuera
de `src/mocks/**` sin pasar por la ruta que la restricción de imports mira.

Los handles se mueven a `onoff/literary-work/literary-work-teasers.mock.ts` y a
`onoff/collection/collections.mock.ts`, junto a las obras y a la prosa de las que derivan. Esa
ruta ya la cubre la restricción vigente, así que no hay regla nueva que escribir ni lista que
mantener: los consumidores legítimos de dentro de `src/mocks/**` la alcanzan igual, y cualquiera
de afuera choca contra el gate `lint`.

Los agregadores quedan con las colecciones, los derivados y los selectores por capacidad.

La contrapartida, aceptada: la vía por nombre deja de cubrirla `lint` y pasa a cubrirla
`typecheck`, porque el símbolo ya no existe en el agregador. Se verificó de las dos formas — un
import por la ruta nueva desde `src/app/**` falla el lint con el mensaje de la restricción, y los
imports por nombre que quedaron colgando rompieron con TS2459.
