+++ src/mocks/onoff-literary-work-teasers.mock.ts
@@ -6,11 +6,6 @@ import type {
 	LiteraryWorkTeaser,
 } from '@models/literary-work.model';
 import type { MediaTeaser } from '@models/media.model';
-import { createLiteraryWorkExcerpt } from '@models/literary-work-excerpt.model';
-import { createMarkdown } from '@models/markdown.model';
-import type { SanitizedHtml } from '@models/sanitized-html.model';
-import { markdownToSanitizedHtml } from '@utils/markdown-pipeline.utils';
-import { embeddedAuthorTeaserMock } from './author.mock';
 import { onoffMediaMock, onoffYouTubeVideosMock, toMediaTeaser } from './onoff-media.mock';
 import { onoffLiteraryWorksMock } from './onoff-literary-works.mock';
 
@@ -31,68 +26,19 @@ function toAuthorTeaser(author: Author): AuthorTeaser {
 		resources: [],
 	};
 }
-import { elOdioLiteraryWorkMock } from './onoff/literary-work/el-odio.literary-work.mock';
-import { elTratadoDeLosPlaceresLiteraryWorkMock } from './onoff/literary-work/el-tratado-de-los-placeres.literary-work.mock';
-import { geometriaLiteraryWorkMock } from './onoff/literary-work/geometria.literary-work.mock';
-import { lasDosAntorchasLiteraryWorkMock } from './onoff/literary-work/las-dos-antorchas.literary-work.mock';
-import { lasEscalerasLiteraryWorkMock } from './onoff/literary-work/las-escaleras.literary-work.mock';
-import { losPeldanosLiteraryWorkMock } from './onoff/literary-work/los-peldanos.literary-work.mock';
-import { neronLiteraryWorkMock } from './onoff/literary-work/neron.literary-work.mock';
-import { palacioNueveFronterasLiteraryWorkMock } from './onoff/literary-work/el-palacio-de-las-nueve-fronteras.literary-work.mock';
-import elOdioMd from './onoff/literary-work/el-odio.md?raw';
-import elTratadoDeLosPlaceresMd from './onoff/literary-work/el-tratado-de-los-placeres.md?raw';
-import geometriaMd from './onoff/literary-work/geometria.md?raw';
-import lasDosAntorchasMd from './onoff/literary-work/las-dos-antorchas.md?raw';
-import lasEscalerasMd from './onoff/literary-work/las-escaleras.md?raw';
-import losPeldanosMd from './onoff/literary-work/los-peldanos.md?raw';
-import neronMd from './onoff/literary-work/neron.md?raw';
-import palacioNueveFronterasMd from './onoff/literary-work/el-palacio-de-las-nueve-fronteras.md?raw';
-
-// Espeja el recorte de la query, para que el corpus de dominio diga lo mismo que produce el ACL. El
-// cruce ACL↔dominio detecta una divergencia solo sobre las formas que el canon tiene: ninguna obra
-// arranca hoy con un renglón en blanco, así que ese caso lo sostiene únicamente que las dos
-// implementaciones se toquen juntas.
-function toExcerptHtml(markdown: string): SanitizedHtml {
-	const blocks = markdown.split('\r\n\r\n')[0].split('\n\n');
-	const firstBlock = blocks.find((block) => block !== '') ?? '';
-	return markdownToSanitizedHtml(createMarkdown(firstBlock));
-}
-
-// Deriva el teaser desde la obra completa: conserva los campos de la vista base, reemplaza los autores
-// por su variante AuthorTeaser y expone el arranque de la sección de apertura como `excerpt`
-// — ver LITERARY_WORK_DESIGN.md §2.
-function toTeaser(literaryWork: LiteraryWork, markdown: string): LiteraryWorkTeaser {
-	return {
-		_id: literaryWork._id,
-		slug: literaryWork.slug,
-		title: literaryWork.title,
-		coverImage: literaryWork.coverImage,
-		totalReadingTime: literaryWork.totalReadingTime,
-		sectionCount: literaryWork.sectionCount,
-		tags: literaryWork.tags,
-		mediaSources: literaryWork.mediaSources.map(toMediaTeaser),
-		authors: [embeddedAuthorTeaserMock],
-		excerpt: createLiteraryWorkExcerpt({
-			title: literaryWork.content[0].title,
-			bodyHtml: toExcerptHtml(markdown),
-		}),
-	};
-}
-
-export const palacioNueveFronterasLiteraryWorkTeaserMock = toTeaser(
-	palacioNueveFronterasLiteraryWorkMock,
-	palacioNueveFronterasMd,
-);
-export const geometriaLiteraryWorkTeaserMock = toTeaser(geometriaLiteraryWorkMock, geometriaMd);
-export const losPeldanosLiteraryWorkTeaserMock = toTeaser(losPeldanosLiteraryWorkMock, losPeldanosMd);
-export const lasEscalerasLiteraryWorkTeaserMock = toTeaser(lasEscalerasLiteraryWorkMock, lasEscalerasMd);
-export const elOdioLiteraryWorkTeaserMock = toTeaser(elOdioLiteraryWorkMock, elOdioMd);
-export const elTratadoDeLosPlaceresLiteraryWorkTeaserMock = toTeaser(
-	elTratadoDeLosPlaceresLiteraryWorkMock,
-	elTratadoDeLosPlaceresMd,
-);
-export const lasDosAntorchasLiteraryWorkTeaserMock = toTeaser(lasDosAntorchasLiteraryWorkMock, lasDosAntorchasMd);
-export const neronLiteraryWorkTeaserMock = toTeaser(neronLiteraryWorkMock, neronMd);
+// Los teasers por obra viven bajo `onoff/literary-work/`, junto a las obras de las que derivan: es la
+// ruta que la restricción de imports alcanza, y este agregador es uno de los consumidores que la
+// convención sí admite.
+import {
+	elOdioLiteraryWorkTeaserMock,
+	elTratadoDeLosPlaceresLiteraryWorkTeaserMock,
+	geometriaLiteraryWorkTeaserMock,
+	lasDosAntorchasLiteraryWorkTeaserMock,
+	lasEscalerasLiteraryWorkTeaserMock,
+	losPeldanosLiteraryWorkTeaserMock,
+	neronLiteraryWorkTeaserMock,
+	palacioNueveFronterasLiteraryWorkTeaserMock,
+} from './onoff/literary-work/literary-work-teasers.mock';
 
 export const onoffLiteraryWorkTeasersMock: LiteraryWorkTeaser[] = [
 	palacioNueveFronterasLiteraryWorkTeaserMock,
